pass resources from CWL -> TES

+++ rabix-backend-local/src/main/java/org/rabix/backend/local/tes/model/TESVolume.java
@@ -8,7 +8,7 @@ public class TESVolume {
   @JsonProperty("name")
   private String name;
   @JsonProperty("sizeGb")
-  private Integer sizeGb;
+  private Double sizeGb;
   @JsonProperty("source")
   private String source;
   @JsonProperty("mountPoint")
@@ -17,7 +17,7 @@ public class TESVolume {
   private boolean readonly;
   
   @JsonCreator
-  public TESVolume(@JsonProperty("name") String name, @JsonProperty("sizeGb") Integer sizeGb, @JsonProperty("source")  String source, @JsonProperty("mountPoint") String mountPoint, @JsonProperty("readonly") boolean readonly) {
+  public TESVolume(@JsonProperty("name") String name, @JsonProperty("sizeGb") Double sizeGb, @JsonProperty("source")  String source, @JsonProperty("mountPoint") String mountPoint, @JsonProperty("readonly") boolean readonly) {
     this.name = name;
     this.sizeGb = sizeGb;
     this.source = source;
@@ -33,11 +33,11 @@ public class TESVolume {
     this.name = name;
   }
 
-  public Integer getSizeGb() {
+  public Double getSizeGb() {
     return sizeGb;
   }
 
-  public void setSizeGb(Integer sizeGb) {
+  public void setSizeGb(Double sizeGb) {
     this.sizeGb = sizeGb;
   }
